apply fixed-point

+++ include/reactphysics3d/collision/narrowphase/GJK/VoronoiSimplex.h
@@ -106,7 +106,7 @@ class VoronoiSimplex {
 
         /// Compute point of a line segment that is closest to the origin
         void computeClosestPointOnSegment(const Vector3& a, const Vector3& b, int& bitUsedVertices,
-                                          float& t) const;
+                                          decimal& t) const;
 
         /// Compute point of a triangle that is closest to the origin
         void computeClosestPointOnTriangle(const Vector3& a, const Vector3& b,
