+++ testbed/common/ConcaveMesh.cpp
@@ -35,9 +35,9 @@ ConcaveMesh::ConcaveMesh(reactphysics3d::BodyType type, bool isSimulationCollide
              mVBOIndices(GL_ELEMENT_ARRAY_BUFFER) {
 
     // Compute the scaling matrix
-    mScalingMatrix = openglframework::Matrix4(scaling.x, 0, 0, 0,
-                                              0, scaling.y, 0, 0,
-                                              0, 0, scaling.z, 0,
+    mScalingMatrix = openglframework::Matrix4((float)scaling.x, 0, 0, 0,
+                                              0, (float)scaling.y, 0, 0,
+                                              0, 0, (float)scaling.z, 0,
                                               0, 0, 0, 1);
